add parameters selection for time, round, y and processing steps
The aim was to be able to skip deskewing in order to only have registration
from Big Sticher using tilted images. To test that,and other things in the
future, it was best to be able to run computation on a subset of data, like
being able to select specific time steps, rounds or y and z positions.