replace RemoteMMCore by CMMCorePlus
The pymmcore-plus github repo states that RemoteMMCore is not thread
safe, that's what may be causing several problems like not properly
closed acquisition preventing new acquisitions.
I just replaced RemoteMMCore by CMMCorePlus, hoping the objects have
the exact same methods.
We'll test acquiring images as it, but we'll proably need to get rid
of the several with clauses and pass the CMMCorePlus object to the
napari widget to control the microscope's components.