remove with clauses and many instantiations
The with clause requires a __enter__ method that was not implemented in the
CMMCorePlus class. Then the microscope configuration was lost during
setup of the laser, probably due to instantiation on a new CMMCorePlus  object.
So we now pass the CMMCorePlus object to the instrument_control_widget
and use only this one to control the lasers and camera.